// libs/chats/src/lib/data/models/message.ts
import { Profile } from '@tt/common';

export type Message = {
  id: number;
  userFromId: number;
  personalChatId: number;
  text: string;
  createdAt: string;
  isRead: boolean;
  updatedAt?: Date;
  isMine: boolean;
  user: Profile;
};

export type SendMessageResponse = {
  id: number;
  userFromId: number;
  personalChatId: number;
  text: string;
  createdAt: string;
  isRead: boolean;
  updatedAt: string;
};